+++ dvha/main.py
@@ -962,8 +962,9 @@ class MainApp(wx.App):
 
         initialize_directories_and_settings()
         if is_windows():
-            from dvha.tools.windows_reg_edit import set_ie_emulation_level
+            from dvha.tools.windows_reg_edit import set_ie_emulation_level, set_ie_lockdown_level
             set_ie_emulation_level()
+            set_ie_lockdown_level()
 
         self.SetAppName('DVH Analytics')
         self.frame = DVHAMainFrame(None, wx.ID_ANY, "")