+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.6.8 (TBD)
     variables using backward elimination
  - [Machine Learning] Fix for [issue 29](https://github.com/cutright/DVH-Analytics/issues/29)
  - [DICOM Importer] Fix for [issue 30](https://github.com/cutright/DVH-Analytics/issues/30)
+ - [Misc] LICENSE was renamed to LICENSE.txt in v0.6.7, but MANIFEST.in and paths.LICENSE_PATH were not updated
 
 v0.6.7 (2020.01.13)
 --------------------